Remove system namespace
This moves the contents of the generated system namespace into the
generated main namespace.

Previously, the system namespace would contain quite a few components,
but more and more I'm keeping those components elsewhere in the app
making the namespace little more than a few functions and a lot of
requires.

As we need a main namespace to produce an uberjar I'm tending to keep
the overall system definition next to the main entry point, pulling in
components from elsewhere in the system.